Feat#10 : 주문 기능 구현

+++ src/main/java/com/bootakhae/webapp/product/dto/ProductDto.java
@@ -16,9 +16,9 @@ public class ProductDto {
 
     private String name;
 
-    private Integer price;
+    private Long price;
 
-    private Integer stock;
+    private Long stock;
 
     private String producer;
