+++ src/main/java/com/bootakhae/webapp/product/vo/response/ResponseProduct.java
@@ -11,9 +11,9 @@ public class ResponseProduct {
 
     private String name;
 
-    private Integer price;
+    private Long price;
 
-    private Integer stock;
+    private Long stock;
 
     private String producer;
